Move socket tests to the bottom of the page and pad.
