+++ packages/client/src/App.tsx
@@ -1,14 +1,11 @@
 import { useEffect, useState } from "react";
 import { Routes, Route, Link } from "react-router-dom";
 import { GameList } from "./GameList";
-import { SERVER_URL } from "./requests";
 import "./App.css";
 
-import io from "socket.io-client";
 import { GamePage } from "./GamePage";
 import { GameCreationForm } from "./GameCreationForm";
-
-const socket = io(SERVER_URL);
+import { SocketTest } from "./sockets";
 
 function App() {
   return (
@@ -23,33 +20,6 @@ function App() {
 }
 
 function Home() {
-  const [isConnected, setIsConnected] = useState(socket.connected);
-  const [lastPong, setLastPong] = useState<string | null>(null);
-
-  useEffect(() => {
-    socket.on("connect", () => {
-      setIsConnected(true);
-    });
-
-    socket.on("disconnect", () => {
-      setIsConnected(false);
-    });
-
-    socket.on("pong", () => {
-      setLastPong(new Date().toISOString());
-    });
-
-    return () => {
-      socket.off("connect");
-      socket.off("disconnect");
-      socket.off("pong");
-    };
-  }, []);
-
-  const sendPing = () => {
-    socket.emit("ping");
-  };
-
   return (
     <>
       <main>
@@ -57,12 +27,9 @@ function Home() {
         <div>
           <GameCreationForm />
           <hr />
-          <h3>socket.io tests...</h3>
-          <p>Connected: {"" + isConnected}</p>
-          <p>{`Last pong: ${lastPong || "-"}`}</p>
-          <button onClick={sendPing}>Send ping</button>
-          <hr />
           <GameList />
+          <hr />
+          <SocketTest />
           <hr />
         </div>
       </main>
